Replace dynamic autolaod with static map
The old autoload iterates recursively over all filenames and tries to
dynamically map classnames to them. This is cool, but getting the
classname out of the filenames is kind of an intense excercise in string
manipulation.

This replaces that with a static associative array that maps classnames
to filepaths, which seems to which just fine.