Remove typehint from getTrackingsByOrderNo
This removes the ShipwireItems typehint from getTrackingsByOrderNo, because the
method returns null if the response from Shipwire is empty. This violates the typehint
and causes an error whenever the Shipwire request fails.

Perhaps it'd be better to throw an exception in this case, but for now this solves the type
problem.