Move request error checking into api method
Each webhook method had the same if condition that checked whether or
not the request was successful and threw a ShipwireException with the
message of the bad HTTP response.

None of the functionality was specific to the top level methods, since
we most likely want to report an error closer to where the request
finishes.

This also DRYs up the methods a lot.