Fix bug in ShipwireResponse where the response doesn't have a body
Some endpoints (like deleteSecret) do not send a body back with the
response, since the HTTP status code should be good enough info for
whether or not it was successful.

If the response doesn't hvae a body, set the `_resource` field to be an
empty `ShipmentResource` to make the type system happy.