Add makePagination method to Shipwire
A previous commit added generator functionality to Shipwire->orders,
because the SDK didn't support it. Now that we're adding the same
pagination support to the Shipwire->stock method, we might as well make
a generic function that runs through Shipwire API pagination instead of
reinventing the wheel for each method.